Screens Main, CreateGame
Dependencies for MVVM has been added and buildFeatures for data binding has been set to true. Two screens - Main and CreateGame - has been created with Activity, ViewModel and ViewModelFactory. Main serves as a Main Menu and has New Game button now. By clicking on this button, CreateGameActivity is launched. This activity will serve for game setup: adding players for the game on a listView and perhaps score to reach to win the game. After the setup, GameActivity will be launched to track the score, this will be added in the future.