Score tracking
Tracking score on GameActivity has been completed and is working now. Score of current player is highlighted by background color (of the player 'item'). For each player, score of 3 throws and its multipliers is edited. Player to reach 0 is winner (winner activity? announcement? in the future). Also, background color is of type Array<MutableLiveData<Int>>, other Mutables will be grouped to arrays as well in the future for cleaner code.